+++ packages/field/src/components/Percent/index.tsx
@@ -1,6 +1,6 @@
 import { useIntl } from '@ant-design/pro-provider';
 import { InputNumber } from 'antd';
-import toNumber from 'lodash.tonumber';
+import { toNumber } from 'lodash-es';
 import type { ReactNode } from 'react';
 import React, { Fragment, useMemo } from 'react';
 import type { ProFieldFC } from '../../PureProField';