+++ packages/form/src/components/index.ts
@@ -63,6 +63,7 @@ export type { ProFormSwitchProps } from './Switch';
 export { default as ProFormText } from './Text';
 export { default as ProFormTextArea } from './TextArea';
 export { default as ProFormTreeSelect } from './TreeSelect';
+export type { ProFormTreeSelectProps } from './TreeSelect';
 export { default as ProFormUploadButton } from './UploadButton';
 export type { ProFormUploadButtonProps } from './UploadButton';
 export { default as ProFormUploadDragger } from './UploadDragger';