+++ packages/table/src/Table.tsx
@@ -62,6 +62,7 @@ function TableRender<T extends Record<string, any>, U, ValueType>(
     defaultClassName: string;
     tableColumn: any[];
     toolbarDom: JSX.Element | null;
+    hideToolbar: boolean;
     searchNode: JSX.Element | null;
     alertDom: JSX.Element | null;
     isLightFilter: boolean;
@@ -84,6 +85,7 @@ function TableRender<T extends Record<string, any>, U, ValueType>(
     defaultSize,
     tableStyle,
     toolbarDom,
+    hideToolbar,
     searchNode,
     style,
     cardProps: propsCardProps,
@@ -329,6 +331,12 @@ function TableRender<T extends Record<string, any>, U, ValueType>(
     if (propsCardProps === false || notNeedCardDom === true || !!props.name)
       return {};
 
+    if (hideToolbar) {
+      return {
+        padding: 0,
+      };
+    }
+
     if (toolbarDom) {
       return {
         paddingBlockStart: 0,
@@ -343,7 +351,14 @@ function TableRender<T extends Record<string, any>, U, ValueType>(
     return {
       padding: 0,
     };
-  }, [notNeedCardDom, pagination, props.name, propsCardProps, toolbarDom]);
+  }, [
+    notNeedCardDom,
+    pagination,
+    props.name,
+    propsCardProps,
+    toolbarDom,
+    hideToolbar,
+  ]);
 
   /** Table 区域的 dom，为了方便 render */
   const tableAreaDom =
@@ -664,8 +679,9 @@ const ProTable = <
     // request 存在且params不为空，且已经请求过数据才需要设置。
     if (
       props.request &&
-      params &&
+      !isEmpty(params) &&
       action.dataSource &&
+      !isEqual(action.dataSource, defaultData) &&
       action?.pageInfo?.current !== 1
     ) {
       action.setPageInfo({
@@ -904,18 +920,22 @@ const ProTable = <
     [action.dataSource, selectedRowKeys],
   ) as T[];
 
+  const hideToolbar = useMemo(
+    () =>
+      options === false &&
+      !headerTitle &&
+      !toolBarRender &&
+      !toolbar &&
+      !isLightFilter,
+    [options, headerTitle, toolBarRender, toolbar, isLightFilter],
+  );
+
   /** 内置的工具栏 */
   const toolbarDom =
     toolBarRender === false ? null : (
       <Toolbar<T>
         headerTitle={headerTitle}
-        hideToolbar={
-          options === false &&
-          !headerTitle &&
-          !toolBarRender &&
-          !toolbar &&
-          !isLightFilter
-        }
+        hideToolbar={hideToolbar}
         selectedRows={selectedRows}
         selectedRowKeys={selectedRowKeys!}
         tableColumn={tableColumn}
@@ -963,6 +983,7 @@ const ProTable = <
       action={action}
       alertDom={alertDom}
       toolbarDom={toolbarDom}
+      hideToolbar={hideToolbar}
       onSortChange={(sortConfig) => {
         if (proSort === sortConfig) return;
         setProSort(sortConfig ?? {});