+++ docs/changelog.md
@@ -1,3 +1,22 @@
+## @ant-design/pro-components@2.7.16
+
+`2024-09-05`
+
+- feat(layout): add menuTextRender props. [1d65e22](https://github.com/ant-design/pro-components/commit/1d65e22)
+- fix(form): props title is not assignable to type ReactNode (#8682). [#8682](https://github.com/ant-design/pro-components/pull/#8682) [@jiAng](https://github.com/jiAng)
+- feat(Form): FormRef support forward nativeElement (#8632). [#8632](https://github.com/ant-design/pro-components/pull/#8632) [@红](https://github.com/红)
+- fix(field): keyboard events of search input in LightSelect (#8629). [#8629](https://github.com/ant-design/pro-components/pull/#8629) [@Anzimu](https://github.com/Anzimu)
+
+## @ant-design/pro-components@2.7.15
+
+`2024-08-05`
+
+- fix(form): pass the OptionType type of ProFormSelect to the inner Select (#8568). [#8568](https://github.com/ant-design/pro-components/pull/#8568) [@dreammaker7](https://github.com/dreammaker7)
+- fix(form): correct line height in list action (#8584). [#8584](https://github.com/ant-design/pro-components/pull/#8584) [@drizzlesconsin](https://github.com/drizzlesconsin)
+- feat(form): export ProFormTreeSelectProps (#8589). [#8589](https://github.com/ant-design/pro-components/pull/#8589) [@fnoopv](https://github.com/fnoopv)
+- fix(table): don't spread key to children (#8588). [#8588](https://github.com/ant-design/pro-components/pull/#8588) [@fnoopv](https://github.com/fnoopv)
+- feat(field): add localizedFormat plugin to dayjs (#8582). [#8582](https://github.com/ant-design/pro-components/pull/#8582) [@lynette-li](https://github.com/lynette-li)
+
 ## @ant-design/pro-components@2.7.12
 
 `2024-07-15`