+++ packages/descriptions/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.53](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@2.5.52...@ant-design/pro-descriptions@2.5.53) (2024-09-12)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [2.5.52](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@2.5.51...@ant-design/pro-descriptions@2.5.52) (2024-09-12)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [2.5.51](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@2.5.50...@ant-design/pro-descriptions@2.5.51) (2024-09-05)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [2.5.50](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@2.5.49...@ant-design/pro-descriptions@2.5.50) (2024-08-05)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
 ## [2.5.49](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@2.5.48...@ant-design/pro-descriptions@2.5.49) (2024-07-18)
 
 **Note:** Version bump only for package @ant-design/pro-descriptions